refactor(dtos): create folders to isolate dto files
