+++ src/core/shared/dto/auth/signIn.dto.ts
@@ -1,9 +1,9 @@
 import { ApiProperty } from '@nestjs/swagger';
 
 export class SignInDto {
-	@ApiProperty()
-	email!: string;
+  @ApiProperty()
+  email!: string;
 
-	@ApiProperty()
-	password!: string;
+  @ApiProperty()
+  password!: string;
 }
