+++ src/core/shared/dto/user/edit-profile.dto.ts
@@ -1,11 +1,11 @@
 import { IsEmail, IsOptional, IsString } from 'class-validator';
 
 export class EditProfileDto {
-	@IsOptional()
-	@IsString()
-	name?: string;
+  @IsOptional()
+  @IsString()
+  name?: string;
 
-	@IsOptional()
-	@IsEmail()
-	email?: string;
+  @IsOptional()
+  @IsEmail()
+  email?: string;
 }
